Backend_Stage_3/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12.0
drf-yasg>=1.20.0
pytest>=6.2.0
pytest-django>=4.4.0

